third_party/blink/web_tests/external/wpt/css/css-values/minmax-length-percent-serialize-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] 'calc(min(10% + 1px) + max(1em + 10%) + min(10% + 20px))' as a specified value should serialize as 'calc(30% + 1em + 21px)'.
  assert_equals: 'calc(30% + 1em + 21px)' should round-trip exactly in specified values. expected "calc(30% + 1em + 21px)" but got "calc((30% + 1em) + 21px)"
[FAIL] 'calc(1em + max(10% + 20px) + 5% + min(1em + 10%) + 10px)' as a specified value should serialize as 'calc(25% + 2em + 30px)'.
  assert_equals: 'calc(25% + 2em + 30px)' should round-trip exactly in specified values. expected "calc(25% + 2em + 30px)" but got "calc((25% + 2em) + 30px)"
Harness: the test ran to completion.
